整理 LogMessages
