--- TeamCode/src/main/java/org/betastudio/ftc/message/LogMessage.java
@@ -1,5 +1,0 @@
-package org.betastudio.ftc.message;
-
-public abstract class LogMessage implements Message {
-	public abstract TelemetryMsg buildTelemetryMsg();
-}
