--- TeamCode/src/main/java/org/betastudio/ftc/message/StringMsg.java
@@ -1,24 +1,0 @@
-package org.betastudio.ftc.message;
-
-import androidx.annotation.NonNull;
-
-import org.betastudio.ftc.ui.telemetry.TelemetryLine;
-
-public class StringMsg extends LogMessage {
-	private final String message;
-
-	public StringMsg(final String message) {
-		this.message = message;
-	}
-
-	@NonNull
-	@Override
-	public String toString() {
-		return message;
-	}
-
-	@Override
-	public TelemetryMsg buildTelemetryMsg() {
-		return new TelemetryMsg(new TelemetryLine(message));
-	}
-}
